+++ src/RequestForm.js
@@ -44,6 +44,7 @@ import {
   Timepicker,
   Checkbox,
   AccordionStatus,
+  NoValue,
 } from '@folio/stripes/components';
 import stripesFinalForm from '@folio/stripes/final-form';
 
@@ -1580,7 +1581,9 @@ class RequestForm extends React.Component {
                             {isEditForm &&
                               <KeyValue
                                 label={<FormattedMessage id="ui-requests.status" />}
-                                value={<FormattedMessage id={requestStatusesTranslations[request.status]} />}
+                                value={(requestStatusesTranslations[request.status]
+                                  ? <FormattedMessage id={requestStatusesTranslations[request.status]} />
+                                  : <NoValue />)}
                               />}
                           </Col>
                           <Col xs={2}>